+++ docs/make.jl
@@ -1,4 +1,5 @@
 using Documenter, Aqua, Changelog
+using DocumenterInterLinks: InterLinks
 
 # Generate a Documenter-friendly changelog from CHANGELOG.md
 Changelog.generate(
@@ -6,6 +7,10 @@ Changelog.generate(
     joinpath(@__DIR__, "..", "CHANGELOG.md"),
     joinpath(@__DIR__, "src", "release-notes.md");
     repo = "JuliaTesting/Aqua.jl",
+)
+
+links = InterLinks(
+    "ExplicitImports" => "https://juliatesting.github.io/ExplicitImports.jl/stable/",
 )
 
 makedocs(;
@@ -30,9 +35,11 @@ makedocs(;
             "piracies.md",
             "persistent_tasks.md",
             "undocumented_names.md",
+            "explicit_imports.md",
         ],
         "release-notes.md",
     ],
+    plugins = [links],
 )
 
 deploydocs(; repo = "github.com/JuliaTesting/Aqua.jl", push_preview = true)